feat: remove species m:n table
